requirements.txt
openai

python-dotenv

google-generativeai

langchain_google_genai

langchain_ollama

langchain_groq

langchain